+++ example/example.dart
@@ -1,25 +1,64 @@
+import 'dart:io';
+
+import 'package:args/args.dart';
 import 'package:bishop/bishop.dart';
 
 import 'play.dart';
 
+final parser = ArgParser()
+  ..addFlag('ai', abbr: 'a', negatable: false)
+  ..addOption('variant', abbr: 'v', defaultsTo: 'chess')
+  ..addOption('pgn', abbr: 'p', help: 'PGN output file')
+  ..addFlag('help', abbr: 'h', negatable: false);
+
 void main(List<String> args) async {
-  String v = args.isNotEmpty ? args.first : 'standard';
+  final parsedArgs = parser.parse(args);
+  if (parsedArgs['help']) {
+    print(parser.usage);
+    return;
+  }
+  String v = parsedArgs['variant'];
+  bool ai = parsedArgs['ai'];
   Variant variant = variantFromString(v) ?? Variant.standard();
   print('Starting game with variant ${variant.name}');
   await Future.delayed(const Duration(seconds: 3));
   Game game = Game(variant: variant);
+  Engine engine = Engine(game: game);
   int i = 0;
 
   while (!game.gameOver) {
+    String playerName = Bishop.playerName[game.turn];
     print(game.ascii());
     print(game.fen);
-    Move m = game.getRandomMove();
-    print('${Bishop.playerName[game.turn]}: ${game.toSan(m)}');
+    Move? m;
+    if (ai) {
+      printYellow('~~ $playerName is thinking..');
+      EngineResult res = await engine.search();
+      printYellow('Best Move: ${formatEngineResult(res, game)}');
+      m = res.move;
+    } else {
+      m = game.getRandomMove();
+    }
+    if (m == null) {
+      printRed('couldn\'t find a move');
+    }
+    print('$playerName: ${game.toSan(m!)}');
     game.makeMove(m);
     i++;
     if (i >= 200) break;
   }
   print(game.ascii());
-  print(game.pgn());
+  printYellow(game.pgn());
   printCyan(game.result?.readable ?? 'Game Over (too long)');
+  if (parsedArgs['pgn'] != null) {
+    final f = File(parsedArgs['pgn']!);
+    f.writeAsStringSync(game.pgn(includeVariant: true));
+    printMagenta('Wrote PGN to ${parsedArgs['pgn']}');
+  }
+}
+
+String formatEngineResult(EngineResult res, Game game) {
+  if (!res.hasMove) return 'No Move';
+  String san = game.toSan(res.move!);
+  return '$san (${res.eval}) [depth ${res.depth}]';
 }