+++ lib/src/utils.dart
@@ -38,19 +38,8 @@ bool validateFen({
 
 /// Looks up a built in variant by name.
 /// If [allowIncomplete], partial matches are allowed.
-Variant? variantFromString(String name, {bool allowIncomplete = true}) {
-  Variants? v = Variants.values.firstWhereOrNull(
-    (e) => e.name.toLowerCase() == name.toLowerCase().replaceAll(' ', ''),
-  );
-  if (v == null && allowIncomplete) {
-    final vs = Variants.values
-        .where((e) => e.name.toLowerCase().startsWith(name.toLowerCase()));
-    if (vs.length == 1) {
-      v = vs.first;
-    }
-  }
-  return v?.build();
-}
+Variant? variantFromString(String name, {bool allowIncomplete = true}) =>
+    Variants.match(name, allowIncomplete: allowIncomplete)?.build();
 
 // Variant? variantFromString(String name, {bool allowIncomplete = true}) =>
 //     (Variants.values.firstWhereOrNull(